Recalibrate voltage sensor for new voltage divider.
The voltage divider is now 22.1 kohm and 100kohm resistors, so recalibrate. In addition,
update the README with documentation, and add new red and yellow LEDs. Now, once init is
completed, the green LED lights and stays on. Then, the yellow LED flashes for each sample,
and red LED turns on if a sensor generates an error.